Move infrastructure modules into core package
