+++ pyproject.toml
@@ -43,12 +43,17 @@ zapzap = "zapzap.__main__:main"
 packages = [
   "zapzap",
   "zapzap.app",
-  "zapzap.config",
   "zapzap.controllers",
   "zapzap.controllers.settings",
+  "zapzap.core",
+  "zapzap.core.config",
+  "zapzap.core.diagnostics",
+  "zapzap.core.environment",
+  "zapzap.core.i18n",
+  "zapzap.core.platform",
+  "zapzap.core.theme",
   "zapzap.models",
   "zapzap.models.settings",
-  "zapzap.platform",
   "zapzap.resources",
   "zapzap.services",
   "zapzap.views",
@@ -56,7 +61,6 @@ packages = [
   "zapzap.views.browser",
   "zapzap.views.components",
   "zapzap.views.settings_components",
-  "zapzap.debug",
   "zapzap.notifications",
   "zapzap.webengine"
 ]
