+++ zapzap/core/__init__.py
@@ -1,0 +1,1 @@
+"""Core infrastructure for ZapZap."""
