+++ zapzap/core/config/__init__.py
@@ -1,0 +1,1 @@
+"""Configuration and persistence infrastructure."""
