+++ zapzap/core/diagnostics/RuntimeEnvironmentDebug.py
@@ -7,7 +7,7 @@ from typing import Dict, Any
 from datetime import datetime
 from PyQt6.QtCore import PYQT_VERSION_STR, QT_VERSION_STR
 from zapzap import __appname__, __version__
-from zapzap.services.EnvironmentDetector import EnvironmentDetector
+from zapzap.core.environment.EnvironmentDetector import EnvironmentDetector
 
 
 class RuntimeEnvironmentDebug:
