+++ zapzap/core/diagnostics/__init__.py
@@ -1,4 +1,4 @@
-from zapzap.debug.CrashDumpHandler import CrashDumpHandler
+from zapzap.core.diagnostics.CrashDumpHandler import CrashDumpHandler
 from zapzap import __appname__
 
 crash_handler = CrashDumpHandler(
