+++ zapzap/core/environment/EnvironmentDetector.py
