+++ zapzap/core/i18n/__init__.py
@@ -1,0 +1,1 @@
+"""Internationalization infrastructure."""
