+++ zapzap/core/theme/ThemeManager.py
@@ -14,8 +14,8 @@ from PyQt6.QtGui import QPalette
 from PyQt6.QtWidgets import QApplication
 
 from zapzap.resources.theme_stylesheet import ThemeStylesheet
-from zapzap.services.SystemThemeMonitor import SystemThemeMonitor
-from zapzap.services.SettingsManager import SettingsManager
+from zapzap.core.theme.SystemThemeMonitor import SystemThemeMonitor
+from zapzap.core.config.SettingsManager import SettingsManager
 
 
 class ThemeManager(QObject):
