+++ zapzap/models/settings/about_settings_model.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from PyQt6.QtCore import PYQT_VERSION_STR, QT_VERSION_STR
 
 from zapzap import __appname__, __bugreport__, __donationPage__, __version__, __website__
-from zapzap.services.EnvironmentDetector import EnvironmentDetector
+from zapzap.core.environment.EnvironmentDetector import EnvironmentDetector
 
 
 class AboutSettingsModel:
