+++ zapzap/models/settings/performance_experimental_settings_model.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from zapzap.services.SettingsManager import SettingsManager
+from zapzap.core.config.SettingsManager import SettingsManager
 
 
 class PerformanceExperimentalSettingsModel:
