+++ zapzap/models/settings/system_startup_settings_model.py
@@ -3,9 +3,9 @@
 from zapzap.services.AutostartManager import AutostartManager
 from zapzap.services.DictionariesManager import DictionariesManager
 from zapzap.services.DownloadManager import DownloadManager
-from zapzap.services.SettingsManager import SettingsManager
-from zapzap.services.SetupManager import SetupManager
-from zapzap.services.TranslationManager import TranslationManager
+from zapzap.core.config.SettingsManager import SettingsManager
+from zapzap.core.environment.SetupManager import SetupManager
+from zapzap.core.i18n.TranslationManager import TranslationManager
 
 
 class SystemStartupSettingsModel:
