+++ zapzap/notifications/FreedesktopNotificationBackend.py
@@ -10,7 +10,7 @@ from PyQt6.QtWebEngineCore import QWebEngineNotification
 
 from zapzap.webengine import WebView
 from zapzap.resources.TrayIcon import TrayIcon
-from zapzap.services.SettingsManager import SettingsManager
+from zapzap.core.config.SettingsManager import SettingsManager
 from zapzap import __appname__
 
 # -----------------------------------------------------------------------------
