+++ zapzap/services/CssPreviewService.py
@@ -6,7 +6,7 @@ import urllib.request
 
 from PyQt6.QtCore import QStandardPaths
 
-from zapzap.services.SettingsManager import SettingsManager
+from zapzap.core.config.SettingsManager import SettingsManager
 
 
 class CssPreviewService:
